requirements.txt
flask
flask-mysqldb
mysql-connector-python
flask-cors
cryptography
pycryptodome